Relocate engineering utilities from Next.js pages
